Move text-only /percolate to /beta/percolate
